fix powershell for multiple os

+++ build_and_tests.ps1
@@ -2,18 +2,20 @@ param (
  [string]$Configuration = "Debug"
 )
 
-
-
 $homeDirectory = $Env:HOME
+
+if (!$homeDirectory) {
+    $homeDirectory = $Env:UserProfile
+}
 
 $nugetPackageCacheDirectory = Join-Path $homeDirectory ".nuget" "packages" "MSBuild.AdditionalTasks"
 
 Write-Host $nugetPackageCacheDirectory
 
-Get-ChildItem -Path $nugetPackageCacheDirectory -Recurse -Hidden | Remove-Item -Force
-
-Remove-Item -Path $nugetPackageCacheDirectory -Force -Recurse
-
+if (!(Test-Path $nugetPackageCacheDirectory)) {
+    Get-ChildItem -Path $nugetPackageCacheDirectory -Recurse -Hidden | Remove-Item -Force
+    Remove-Item -Path $nugetPackageCacheDirectory -Force -Recurse
+}
 
 & dotnet build ./src/MSBuild.AdditionalTasks.sln /binaryLogger:msbuild.binlog /nodeReuse:false -c $Configuration
